refactor: Moved stories into their own section
